// assets/BHY_Framework/UI/Enums/UILayerType.ts
/**
 * UI层级类型枚举
 * 定义界面显示的层级顺序
 */
export enum UILayerType {
    /** 底层界面 */
    BACKGROUND_0 = 0,
    /** 正常界面 */
    NORMAL_1 = 100,
    /** 弹窗界面 */
    POPUP_2 = 200,
    /** 提示界面 */
    TIPS_3 = 300,
    /** 最上层界面 */
    TOP_4 = 400,
    /** 系统界面 */
    SYSTEM_5 = 500,
    /** 系统界面 */
    SYSTEM_6 = 600
} 